feat: add life cycle example
